fyp/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pycparser==2.21
pytz==2023.3
sqlparse==0.4.3
